move fatal message handling into MainWindow
