+++ ui/mainwindow.h
@@ -50,6 +50,10 @@ class MainWindow : public QMainWindow
     void quitHost();
     void detachProbe();
 
+    void fatalMessageReceived(const QString &app, const QString &message,
+                              const QTime &time, const QStringList &backtrace);
+    void copyToClipboard(const QString &message);
+
   private:
     QWidget* createErrorPage(const QModelIndex &index);
