+++ ui/tools/messagehandler/messagehandlerwidget.cpp
@@ -26,37 +26,16 @@
 
 #include <common/endpoint.h>
 #include <common/objectbroker.h>
-#include "messagehandlerclient.h"
 
 #include <QSortFilterProxyModel>
-#include <QDialog>
-#include <QDialogButtonBox>
-#include <QLabel>
-#include <QListWidget>
-#include <QTime>
-#include <QPushButton>
-#include <QClipboard>
-#include <QApplication>
-#include <QSignalMapper>
 #include <QTreeView>
 
 using namespace GammaRay;
-
-static QObject *createClientMessageHandler(const QString &/*name*/, QObject *parent)
-{
-  return new MessageHandlerClient(parent);
-}
 
 MessageHandlerWidget::MessageHandlerWidget(QWidget *parent)
   : QWidget(parent),
     ui(new Ui::MessageHandlerWidget)
 {
-  ObjectBroker::registerClientObjectFactoryCallback<MessageHandlerInterface*>(createClientMessageHandler);
-  MessageHandlerInterface *handler = ObjectBroker::object<MessageHandlerInterface*>();
-
-  connect(handler, SIGNAL(fatalMessageReceived(QString,QString,QTime,QStringList)),
-          this, SLOT(fatalMessageReceived(QString,QString,QTime,QStringList)));
-
   ui->setupUi(this);
 
   proxy = new QSortFilterProxyModel(this);
@@ -77,66 +56,7 @@ MessageHandlerWidget::~MessageHandlerWidget()
 {
 }
 
-void MessageHandlerWidget::fatalMessageReceived(const QString &app, const QString &message,
-                                                const QTime &time, const QStringList &backtrace)
-{
-  if (Endpoint::isConnected() && !qobject_cast<MessageHandlerClient*>(ObjectBroker::object<MessageHandlerInterface*>())) {
-    // only show on remote side
-    return;
-  }
-  QDialog dlg;
-  dlg.setWindowTitle(QObject::tr("QFatal in %1 at %2").arg(app).arg(time.toString()));
 
-  QGridLayout *layout = new QGridLayout;
-
-  QLabel *iconLabel = new QLabel;
-  QIcon icon = dlg.style()->standardIcon(QStyle::SP_MessageBoxCritical, 0, &dlg);
-  int iconSize = dlg.style()->pixelMetric(QStyle::PM_MessageBoxIconSize, 0, &dlg);
-  iconLabel->setPixmap(icon.pixmap(iconSize, iconSize));
-  iconLabel->setSizePolicy(QSizePolicy::Fixed, QSizePolicy::Fixed);
-  layout->addWidget(iconLabel, 0, 0);
-
-  QLabel *errorLabel = new QLabel;
-  errorLabel->setTextFormat(Qt::PlainText);
-  errorLabel->setWordWrap(true);
-  errorLabel->setText(message);
-  layout->addWidget(errorLabel, 0, 1);
-
-  QDialogButtonBox *buttons = new QDialogButtonBox;
-
-  if (!backtrace.isEmpty()) {
-    QListWidget *backtraceWidget = new QListWidget;
-    foreach (const QString &frame, backtrace) {
-      backtraceWidget->addItem(frame);
-    }
-    layout->addWidget(backtraceWidget, 1, 0, 1, 2);
-
-    QPushButton *copyBacktraceButton = new QPushButton(tr("Copy Backtrace"));
-    buttons->addButton(copyBacktraceButton, QDialogButtonBox::ActionRole);
-
-    QSignalMapper *mapper = new QSignalMapper(this);
-    mapper->setMapping(copyBacktraceButton, backtrace.join(QLatin1String("\n")));
-
-    connect(copyBacktraceButton, SIGNAL(clicked()), mapper, SLOT(map()));
-    connect(mapper, SIGNAL(mapped(QString)), this, SLOT(copyToClipboard(QString)));
-  }
-
-  buttons->addButton(QDialogButtonBox::Close);
-  QObject::connect(buttons, SIGNAL(accepted()),
-                    &dlg, SLOT(accept()));
-  QObject::connect(buttons, SIGNAL(rejected()),
-                    &dlg, SLOT(reject()));
-  layout->addWidget(buttons, 2, 0, 1, 2);
-
-  dlg.setLayout(layout);
-  dlg.adjustSize();
-  dlg.exec();
-}
-
-void MessageHandlerWidget::copyToClipboard(const QString &message)
-{
-  QApplication::clipboard()->setText(message);
-}
 
 void MessageHandlerWidget::selectionChanged(const QModelIndex &current, const QModelIndex &previous)
 {
